requirements.txt
fastapi
passlib[bcrypt]
pydantic
python-jose[cryptography]
SQLAlchemy
uvicorn
psycopg2-binary